tehty 4.23 asti

+++ osa4/controllers/blogs.js
@@ -1,27 +1,16 @@
 const blogRouter = require('express').Router()
-const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
-const User = require('../models/user')
-
+const middleware = require('../utils/middleware')
 
 blogRouter.get('/', async (request, response) => {
   const res = await Blog
-  .find({})
-  .populate('user', { username: 1, name: 1 })
+    .find({})
+    .populate('user', { username: 1, name: 1 })
   response.json(res)
 })
   
-blogRouter.post('/', async (request, response) => {
-  console.log(request.token, "!!!")
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!request.token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
-
-  if(!request.body.hasOwnProperty('title') || !request.body.hasOwnProperty('url')){
-    return response.status(400).json({answer: "url and title required"})
-  }
+blogRouter.post('/',middleware.userExtractor, async (request, response) => {
+  const user = request.user
 
   let blog = new Blog({
     title: request.body.title,
@@ -35,16 +24,26 @@ blogRouter.post('/', async (request, response) => {
     blog.likes = request.body.likes
   }
   
-  const res = await blog.save();
+  const res = await blog.save()
   user.blogs = user.blogs.concat(res._id)
   await user.save()
   response.status(201).json(res)
 })
 
-blogRouter.delete('/:id', async (request, response) => {
+blogRouter.delete('/:id',middleware.userExtractor, async (request, response) => {
   const id = request.params.id
-  await Blog.findByIdAndDelete(id)
-  response.status(203).end()
+  const user = request.user
+  const blog = await Blog.findById(id)
+
+  if ( blog.user.toString() === user.id.toString()){
+    await Blog.findByIdAndDelete(id)
+    response.status(203).end()
+  }
+  else{
+    response.status(400).end()
+  }
+
+
 })
 
 blogRouter.put('/:id', (request, response) => {
@@ -57,12 +56,12 @@ blogRouter.put('/:id', (request, response) => {
     likes: body.likes
   }
 
-  const updatedBlog = Blog.findByIdAndUpdate(request.params.id, blog, { new: true, runValidators: true, context: 'query' })
-  .then(updatedNote => {
-    console.log(updatedNote, '!')
-    console.log(blog)
-    response.json(updatedNote)
-  })
+  Blog.findByIdAndUpdate(request.params.id, blog, { new: true, runValidators: true, context: 'query' })
+    .then(updatedNote => {
+      console.log(updatedNote, '!')
+      console.log(blog)
+      response.json(updatedNote)
+    })
 })
 
 module.exports = blogRouter
